refactor: useQueryData 로직 훅 변경
